+++ src/views/components/index.ts
@@ -1,11 +1,12 @@
-export * from './actions'
 export * from './UI'
+export * from './actions'
 export * from './dropdowns'
 export * from './forms'
 export * from './inputs'
 export * from './timeline'
 export * from './AdminDashboard'
 export * from './CDCChoiceForm'
+export * from './ChoisirCahierDesChargesFormulaire'
 export * from './Component'
 export * from './Container'
 export * from './DateInput'
