+++ src/views/pages/demanderDelaiPage/DemanderDelai.tsx
@@ -11,6 +11,7 @@ import {
   SecondaryLinkButton,
   InfoBox,
   ExternalLink,
+  ChoisirCahierDesChargesFormulaire,
 } from '@components'
 import routes from '@routes'
 import { Project } from '@entities'
@@ -21,8 +22,6 @@ import format from 'date-fns/format'
 
 import { dataId } from '../../../helpers/testId'
 import { hydrateOnClient } from '../../helpers'
-
-import { ChoisirCahierDesChargesFormulaire } from '../choisirCahierDesChargesPage'
 
 type DemanderDelaiProps = {
   request: Request
@@ -43,6 +42,7 @@ export const DemanderDelai = PageLayout((props: DemanderDelaiProps) => {
   const doitChoisirCahierDesCharges =
     project.appelOffre?.choisirNouveauCahierDesCharges &&
     !project.nouvellesRèglesDInstructionChoisies
+
   const [newRulesOptInSelectionné] = useState(project.nouvellesRèglesDInstructionChoisies)
   const nouvelleDateAchèvementMinimale = new Date(project.completionDueOn).setDate(
     new Date(project.completionDueOn).getDate() + 1
@@ -57,132 +57,133 @@ export const DemanderDelai = PageLayout((props: DemanderDelaiProps) => {
           </h3>
         </div>
 
-        <form action={routes.DEMANDE_DELAI_ACTION} method="post" encType="multipart/form-data">
-          <input type="hidden" name="projectId" value={project.id} />
-          <input type="hidden" name="type" value={'delai'} />
-          <div className="form__group">
-            <SuccessErrorBox success={success} error={error} />
-            {doitChoisirCahierDesCharges && (
-              <>
-                <InfoBox
-                  title="Afin d'accéder au formulaire de demande de délai, vous devez d'abord changer le
+        {doitChoisirCahierDesCharges ? (
+          <>
+            <InfoBox
+              title="Afin d'accéder au formulaire de demande de délai, vous devez d'abord changer le
                   cahier des charges à appliquer"
-                  className="mb-5"
-                >
-                  <p className="m-0">
-                    Pour plus d'informations sur les modalités d'instruction veuillez consulter
-                    cette &nbsp;
-                    <ExternalLink href="https://docs.potentiel.beta.gouv.fr/info/guide-dutilisation-potentiel/comment-faire-une-demande-de-modification-ou-informer-le-prefet-dun-changement">
-                      page d'aide
-                    </ExternalLink>
-                    .
-                  </p>
-                </InfoBox>
-                <ChoisirCahierDesChargesFormulaire
-                  cahiersChargesURLs={cahiersChargesURLs}
-                  projet={project}
-                  redirectUrl={routes.DEMANDER_DELAI(project.id)}
-                />
-              </>
-            )}
+              className="mb-5"
+            >
+              <p className="m-0">
+                Pour plus d'informations sur les modalités d'instruction veuillez consulter cette
+                &nbsp;
+                <ExternalLink href="https://docs.potentiel.beta.gouv.fr/info/guide-dutilisation-potentiel/comment-faire-une-demande-de-modification-ou-informer-le-prefet-dun-changement">
+                  page d'aide
+                </ExternalLink>
+                .
+              </p>
+            </InfoBox>
+            <ChoisirCahierDesChargesFormulaire
+              cahiersChargesURLs={cahiersChargesURLs}
+              projet={project}
+              redirectUrl={routes.DEMANDER_DELAI(project.id)}
+            />
+          </>
+        ) : (
+          <form action={routes.DEMANDE_DELAI_ACTION} method="post" encType="multipart/form-data">
+            <input type="hidden" name="projectId" value={project.id} />
+            <input type="hidden" name="type" value={'delai'} />
+            <div className="form__group">
+              <SuccessErrorBox success={success} error={error} />
 
-            {(newRulesOptInSelectionné || !doitChoisirCahierDesCharges) && (
-              <>
-                <FormulaireChampsObligatoireLégende className="text-right" />
-                <div className="mb-1">Concernant le projet:</div>
-                <ProjectInfo project={project} className="mb-3" />
-                <div {...dataId('modificationRequest-demandesInputs')}>
-                  <div className="flex flex-col gap-5">
-                    <div>
-                      <label>Date théorique d'achèvement</label>
-                      <Input
-                        type="text"
-                        disabled
-                        defaultValue={format(project.completionDueOn, 'dd / MM / yyyy')}
-                        style={{ backgroundColor: '#CECECE' }}
-                        {...dataId('modificationRequest-presentServiceDateField')}
-                      />
-                    </div>
-                    <div>
-                      <label htmlFor="dateAchèvementDemandée">
-                        Saisissez la date limite d'achèvement souhaitée <Astérisque />
-                      </label>
-                      <Input
-                        type="date"
-                        name="dateAchèvementDemandée"
-                        id="dateAchèvementDemandée"
-                        min={format(nouvelleDateAchèvementMinimale, 'yyyy-MM-dd')}
-                        defaultValue={dateAchèvementDemandée}
-                        required
-                        aria-required="true"
-                      />
-                    </div>
-                    <div>
-                      <label htmlFor="justification">
-                        Veuillez nous indiquer les raisons qui motivent votre demande
-                        <br />
-                        <span className="italic">
-                          Pour faciliter le traitement de votre demande, veillez à détailler les
-                          raisons ayant conduit à ce besoin de modification (contexte, facteurs
-                          extérieurs, etc.)
-                        </span>
-                      </label>
-                      <TextArea
-                        name="justification"
-                        id="justification"
-                        defaultValue={justification || ''}
-                        {...dataId('modificationRequest-justificationField')}
-                      />
-                    </div>
-                    {!(project.dcrNumeroDossier || project.numeroGestionnaire) ? (
+              {(newRulesOptInSelectionné || !doitChoisirCahierDesCharges) && (
+                <>
+                  <FormulaireChampsObligatoireLégende className="text-right" />
+                  <div className="mb-1">Concernant le projet:</div>
+                  <ProjectInfo project={project} className="mb-3" />
+                  <div {...dataId('modificationRequest-demandesInputs')}>
+                    <div className="flex flex-col gap-5">
                       <div>
-                        <label htmlFor="numeroGestionnaire">
-                          Identifiant gestionnaire de réseau
-                        </label>
-                        <div className="italic">
-                          Cette indication permettra un traitement plus rapide de votre demande.{' '}
-                          <a href="https://docs.potentiel.beta.gouv.fr/info/guide-dutilisation-potentiel/comment-transmettre-ma-demande-complete-de-raccordement-dcr">
-                            Où trouver mon numéro ?
-                          </a>
-                        </div>
-                        <input
+                        <label>Date théorique d'achèvement</label>
+                        <Input
                           type="text"
-                          name="numeroGestionnaire"
-                          {...dataId('modificationRequest-numeroGestionnaireField')}
-                          id="numeroGestionnaire"
+                          disabled
+                          defaultValue={format(project.completionDueOn, 'dd / MM / yyyy')}
+                          style={{ backgroundColor: '#CECECE' }}
+                          {...dataId('modificationRequest-presentServiceDateField')}
                         />
                       </div>
-                    ) : null}
-                    <div>
-                      <label htmlFor="file">Pièce justificative (si nécessaire)</label>
-                      <Input
-                        type="file"
-                        name="file"
-                        {...dataId('modificationRequest-fileField')}
-                        id="file"
-                      />
+                      <div>
+                        <label htmlFor="dateAchèvementDemandée">
+                          Saisissez la date limite d'achèvement souhaitée <Astérisque />
+                        </label>
+                        <Input
+                          type="date"
+                          name="dateAchèvementDemandée"
+                          id="dateAchèvementDemandée"
+                          min={format(nouvelleDateAchèvementMinimale, 'yyyy-MM-dd')}
+                          defaultValue={dateAchèvementDemandée}
+                          required
+                          aria-required="true"
+                        />
+                      </div>
+                      <div>
+                        <label htmlFor="justification">
+                          Veuillez nous indiquer les raisons qui motivent votre demande
+                          <br />
+                          <span className="italic">
+                            Pour faciliter le traitement de votre demande, veillez à détailler les
+                            raisons ayant conduit à ce besoin de modification (contexte, facteurs
+                            extérieurs, etc.)
+                          </span>
+                        </label>
+                        <TextArea
+                          name="justification"
+                          id="justification"
+                          defaultValue={justification || ''}
+                          {...dataId('modificationRequest-justificationField')}
+                        />
+                      </div>
+                      {!(project.dcrNumeroDossier || project.numeroGestionnaire) ? (
+                        <div>
+                          <label htmlFor="numeroGestionnaire">
+                            Identifiant gestionnaire de réseau
+                          </label>
+                          <div className="italic">
+                            Cette indication permettra un traitement plus rapide de votre demande.{' '}
+                            <a href="https://docs.potentiel.beta.gouv.fr/info/guide-dutilisation-potentiel/comment-transmettre-ma-demande-complete-de-raccordement-dcr">
+                              Où trouver mon numéro ?
+                            </a>
+                          </div>
+                          <input
+                            type="text"
+                            name="numeroGestionnaire"
+                            {...dataId('modificationRequest-numeroGestionnaireField')}
+                            id="numeroGestionnaire"
+                          />
+                        </div>
+                      ) : null}
+                      <div>
+                        <label htmlFor="file">Pièce justificative (si nécessaire)</label>
+                        <Input
+                          type="file"
+                          name="file"
+                          {...dataId('modificationRequest-fileField')}
+                          id="file"
+                        />
+                      </div>
                     </div>
-                  </div>
 
-                  <Button
-                    type="submit"
-                    id="submit"
-                    {...dataId('submit-button')}
-                    className="mt-4 mr-2"
-                  >
-                    Envoyer
-                  </Button>
-                  <SecondaryLinkButton
-                    {...dataId('cancel-button')}
-                    href={routes.USER_LIST_PROJECTS}
-                  >
-                    Annuler
-                  </SecondaryLinkButton>
-                </div>
-              </>
-            )}
-          </div>
-        </form>
+                    <Button
+                      type="submit"
+                      id="submit"
+                      {...dataId('submit-button')}
+                      className="mt-4 mr-2"
+                    >
+                      Envoyer
+                    </Button>
+                    <SecondaryLinkButton
+                      {...dataId('cancel-button')}
+                      href={routes.USER_LIST_PROJECTS}
+                    >
+                      Annuler
+                    </SecondaryLinkButton>
+                  </div>
+                </>
+              )}
+            </div>
+          </form>
+        )}
       </div>
     </UserDashboard>
   )
